handle responseWriter being written to after closure.
fixes #2

If macaron.Recovery is being used and a panic is encountered,
the responseWriter handler will complete without ever writing anything
triggering the `defer grw.close()` to be called.  However,
macaron.Recovery will then continue to use the responseWriter
for writing out the 500 error and stack trace.  To handle this case
we need to track if close() has been called, and if so then we should
just write to the raw, uncompressed output stream.